Add Debug.ts
A module which defines the dbg method as either an NOP function or a
function that uses the `npm debug` to output the string parameter if
Debug.debugging is true and the environment variable DEBUG=DBG.

So looks like a NOP function for dbg is not faster than just using
`npm debug` directly. So as it is now test-nn.c run 1,000,000 in 1.40s
and test-nn.ts is taking 158s so 100x slower.

Finally, the TS version error is wacky:
```
Epoch=900000: error=1142.3513784677418
```
Where as the C version is 0:
```
Epoch=1000000 Error=0.000000
```